Use float to support sub-second precision

+++ src/humanize/time.py
@@ -104,7 +104,7 @@ def naturaldelta(
     Returns:
         str (str or `value`): A natural representation of the amount of time
             elapsed unless `value` is not datetime.timedelta or cannot be
-            converted to int. In that case, a `value` is returned unchanged.
+            converted to float. In that case, a `value` is returned unchanged.
 
     Raises:
         OverflowError: If `value` is too large to convert to datetime.timedelta.
@@ -131,7 +131,7 @@ def naturaldelta(
         delta = value
     else:
         try:
-            value = int(value)
+            value = float(value)
             delta = dt.timedelta(seconds=value)
         except (ValueError, TypeError):
             return str(value)
